feat: add radio button for paid
