--- src/main/java/org/example/answer/answer/controller/AnswerController.java
@@ -1,101 +1,0 @@
-package org.example.answer.answer.controller;
-
-import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotEmpty;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
-import org.example.answer.answer.entity.Answer;
-import org.example.answer.answer.service.AnswerService;
-import org.example.question.question.controller.QuestionController;
-import org.example.question.question.entity.Question;
-import org.example.question.question.service.QuestionService;
-import org.example.user.user.entity.SiteUser;
-import org.example.user.user.service.UserService;
-import org.springframework.http.HttpStatus;
-import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.*;
-import org.springframework.web.server.ResponseStatusException;
-
-import java.security.Principal;
-
-@Controller
-@RequestMapping("/answer")
-@RequiredArgsConstructor
-public class AnswerController {
-    private final QuestionService questionService;
-    private final AnswerService answerService;
-    private final UserService userService;
-
-    @Getter
-    @Setter
-    public static class AnswerForm {
-        @NotEmpty(message = "댓글 내용은 필수입니다.")
-        private String content;
-    }
-
-    @PreAuthorize("isAuthenticated()")
-    @PostMapping("/create/{id}")
-    public String createAnswer(Model model, @PathVariable("id") Integer id,
-                               @Valid AnswerForm answerForm, BindingResult bindingResult, Principal principal) {
-        Question question = questionService.getQuestion(id);
-        SiteUser siteUser = userService.getUser(principal.getName());
-        if (bindingResult.hasErrors()) {
-            model.addAttribute("question", question);
-            return "question_detail";
-        }
-        Answer answer = answerService.create(question, answerForm.getContent(), siteUser);
-
-        return String.format("redirect:/question/detail/%s#answer_%s", answer.getQuestion().getId(),answer.getId());
-    }
-
-    @PreAuthorize("isAuthenticated()")
-    @GetMapping("/modify/{id}")
-    public String answerModify(AnswerForm answerForm, @PathVariable("id") Integer id, Principal principal) {
-        Answer answer = answerService.getAnswer(id);
-        if (!answer.getAuthor().getUsername().equals(principal.getName())) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정권한이 없습니다.");
-        }
-        answerForm.setContent(answer.getContent());
-        return "answer_form";
-    }
-
-    @PreAuthorize("isAuthenticated()")
-    @PostMapping("/modify/{id}")
-    public String answerModify(@Valid AnswerForm answerForm, BindingResult bindingResult,
-                               Principal principal, @PathVariable("id") Integer id) {
-        if (bindingResult.hasErrors()) {
-            return "answer_form";
-        }
-        Answer answer = answerService.getAnswer(id);
-        if (!answer.getAuthor().getUsername().equals(principal.getName())) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정권한이 없습니다.");
-        }
-        answerService.modify(answer, answerForm.getContent());
-        return String.format("redirect:/question/detail/%s#answer_%s", answer.getQuestion().getId(),answer.getId());
-
-    }
-
-    @PreAuthorize("isAuthenticated()")
-    @GetMapping("/delete/{id}")
-    public String answerDelete(Principal principal, @PathVariable("id") Integer id) {
-        Answer answer = answerService.getAnswer(id);
-        if (!answer.getAuthor().getUsername().equals(principal.getName())) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "삭제 권한이 없습니다.");
-        }
-        answerService.delete(answer);
-        return String.format("redirect:/question/detail/%s", answer.getQuestion().getId());
-    }
-
-    @PreAuthorize("isAuthenticated()")
-    @GetMapping("/vote/{id}")
-    public String answerVote(Principal principal, @PathVariable("id") Integer id) {
-        Answer answer = answerService.getAnswer(id);
-        SiteUser siteUser = userService.getUser(principal.getName());
-        answerService.vote(answer, siteUser);
-        return String.format("redirect:/question/detail/%s#answer_%s", answer.getQuestion().getId(),answer.getId());
-    }
-}
