--- src/main/java/org/example/answer/answer/repository/AnswerRepository.java
@@ -1,9 +1,0 @@
-package org.example.answer.answer.repository;
-
-import org.example.answer.answer.entity.Answer;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AnswerRepository extends JpaRepository<Answer,Integer> {
-}
