+++ src/main/java/org/example/domain/answer/answer/repository/AnswerRepository.java
@@ -1,0 +1,9 @@
+package org.example.domain.answer.answer.repository;
+
+import org.example.domain.answer.answer.entity.Answer;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface AnswerRepository extends JpaRepository<Answer,Integer> {
+}
