+++ src/main/java/org/example/domain/answer/answer/service/AnswerService.java
@@ -1,0 +1,52 @@
+package org.example.domain.answer.answer.service;
+
+import lombok.RequiredArgsConstructor;
+import org.example.domain.answer.answer.entity.Answer;
+import org.example.domain.answer.answer.repository.AnswerRepository;
+import org.example.domain.question.question.entity.Question;
+import org.example.domain.question.question.exception.DataNotFoundException;
+import org.example.domain.user.user.entity.SiteUser;
+import org.springframework.stereotype.Service;
+
+import java.time.LocalDateTime;
+import java.util.Optional;
+
+@Service
+@RequiredArgsConstructor
+public class AnswerService {
+    private final AnswerRepository answerRepository;
+
+    public Answer create(Question question, String content, SiteUser author){
+        Answer answer = new Answer();
+        answer.setContent(content);
+        answer.setCreateDate(LocalDateTime.now());
+        answer.setQuestion(question);
+        answer.setAuthor(author);
+        answerRepository.save(answer);
+        return answer;
+    }
+
+    public Answer getAnswer(Integer id){
+        Optional<Answer> answer = answerRepository.findById(id);
+        if(answer.isPresent()){
+            return answer.get();
+        }else{
+            throw new DataNotFoundException("answer not found");
+        }
+    }
+
+    public void modify(Answer answer, String content) {
+        answer.setContent(content);
+        answer.setModifyDate(LocalDateTime.now());
+        answerRepository.save(answer);
+    }
+
+    public void delete(Answer answer) {
+        answerRepository.delete(answer);
+    }
+
+    public void vote(Answer answer, SiteUser siteUser){
+        answer.getVoter().add(siteUser);
+        answerRepository.save(answer);
+    }
+}
