+++ src/main/java/org/example/domain/question/question/exception/DataNotFoundException.java
@@ -1,0 +1,12 @@
+package org.example.domain.question.question.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value = HttpStatus.NOT_FOUND,reason = "entity not found")
+public class DataNotFoundException extends RuntimeException {
+    private static final long serialVersionUID = 1L;
+    public DataNotFoundException(String message) {
+        super(message);
+    }
+}
