+++ src/main/java/org/example/domain/question/question/repository/QuestionRepository.java
@@ -1,0 +1,15 @@
+package org.example.domain.question.question.repository;
+
+import org.example.domain.question.question.entity.Question;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface QuestionRepository extends JpaRepository<Question,Integer> {
+    Question findBySubject(String subject);
+    Question findBySubjectAndContent(String subject,String content);
+    List<Question> findBySubjectLike(String subject);
+    Page<Question> findAll(Pageable pageable);
+}
