+++ src/main/java/org/example/domain/user/user/controller/UserController.java
@@ -1,0 +1,77 @@
+package org.example.domain.user.user.controller;
+
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.Size;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.Setter;
+import org.example.domain.user.user.service.UserService;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.stereotype.Controller;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequiredArgsConstructor
+@RequestMapping("/user")
+public class UserController {
+    private final UserService userService;
+
+    @Getter
+    @Setter
+    public static class UserCreateForm{
+        @Size(min=3 , max=25)
+        @NotEmpty(message = "사용자 ID는 필수항목입니다.")
+        private String username;
+
+        @NotEmpty(message = "비밀번호는 필수항목입니다.")
+        private String password1;
+
+        @NotEmpty(message = "비밀번호 확인은 필수항복입니다.")
+        private String password2;
+
+        @NotEmpty(message = "이메일은 필수항목입니다.")
+        @Email
+        private String email;
+    }
+
+    @GetMapping("/signup")
+    public String signup(UserCreateForm userCreateForm) {
+        return "signup_form";
+    }
+
+    @PostMapping("/signup")
+    public String signup(@Valid UserCreateForm userCreateForm, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return "signup_form";
+        }
+
+        if (!userCreateForm.getPassword1().equals(userCreateForm.getPassword2())) {
+            bindingResult.rejectValue("password2","passwordInCorrect","2개의 패스워드가 일치하지않습니다.");
+            return "signup_form";
+        }
+
+        try{
+            userService.create(userCreateForm.username, userCreateForm.email, userCreateForm.password1);
+        }catch (DataIntegrityViolationException e){
+            e.printStackTrace();
+            bindingResult.reject("signupFailed","이미 등록된 사용자입니다.");
+            return "signup_form";
+        }catch (Exception e){
+            e.printStackTrace();
+            bindingResult.reject("signupFailed",e.getMessage());
+            return "signup_form";
+        }
+
+        return "redirect:/";
+    }
+
+    @GetMapping("/login")
+    public String login(){
+        return "login_form";
+    }
+}
