+++ src/main/java/org/example/domain/user/user/entity/SiteUser.java
@@ -1,0 +1,24 @@
+package org.example.domain.user.user.entity;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Entity
+public class SiteUser {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(unique = true)
+    private String username;
+
+    private String password;
+
+    @Column(unique = true)
+    private String email;
+
+    private boolean paid; // 유료 회원 여부
+}
