+++ src/main/java/org/example/domain/user/user/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package org.example.domain.user.user.repository;
+
+import org.example.domain.user.user.entity.SiteUser;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<SiteUser, Long> {
+    Optional<SiteUser> findByUsername(String username);
+}
