+++ src/main/java/org/example/domain/user/user/role/UserRole.java
@@ -1,0 +1,16 @@
+package org.example.domain.user.user.role;
+
+import lombok.Getter;
+
+@Getter
+public enum UserRole {
+    ADMIN("ROLE_ADMIN"),
+    USER("ROLE_USER"),
+    PAID("ROLE_PAID");
+
+    private String value;
+
+    UserRole(String value) {
+        this.value = value;
+    }
+}
