+++ src/main/java/org/example/global/initData/NotProd.java
@@ -1,0 +1,34 @@
+package org.example.global.initData;
+
+import lombok.RequiredArgsConstructor;
+import org.example.domain.question.question.service.QuestionService;
+import org.example.domain.user.user.service.UserService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.ApplicationRunner;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Lazy;
+import org.springframework.transaction.annotation.Transactional;
+
+@Configuration
+@RequiredArgsConstructor
+public class NotProd {
+    @Autowired
+    @Lazy
+    private NotProd self;
+
+    private final UserService userService;
+    private final QuestionService questionService;
+
+    @Bean
+    public ApplicationRunner initNotProd(){
+        return args -> {
+            self.work1();
+        };
+    }
+
+    @Transactional
+    public void work1(){
+//        questionService.create();
+    }
+}
