--- src/main/java/org/example/question/question/exception/DataNotFoundException.java
@@ -1,12 +1,0 @@
-package org.example.question.question.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value = HttpStatus.NOT_FOUND,reason = "entity not found")
-public class DataNotFoundException extends RuntimeException {
-    private static final long serialVersionUID = 1L;
-    public DataNotFoundException(String message) {
-        super(message);
-    }
-}
