--- src/main/java/org/example/question/question/service/QuestionService.java
@@ -1,67 +1,0 @@
-package org.example.question.question.service;
-
-import lombok.RequiredArgsConstructor;
-import org.example.question.question.entity.Question;
-import org.example.question.question.exception.DataNotFoundException;
-import org.example.question.question.repository.QuestionRepository;
-import org.example.user.user.entity.SiteUser;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
-import org.springframework.stereotype.Service;
-
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-
-@Service
-@RequiredArgsConstructor
-public class QuestionService {
-    private final QuestionRepository questionRepository;
-
-    public List<Question> getList(){
-        return questionRepository.findAll();
-    }
-
-    public Question getQuestion(Integer id){
-        Optional<Question> question = questionRepository.findById(id);
-        if(question.isPresent()){
-            return question.get();
-        }else{
-            throw new DataNotFoundException("question not found");
-        }
-    }
-
-    public void create(String subject, String content, SiteUser author) {
-        Question question = new Question();
-        question.setSubject(subject);
-        question.setCreateDate(LocalDateTime.now());
-        question.setContent(content);
-        question.setAuthor(author);
-        questionRepository.save(question);
-    }
-    public Page<Question> getList(int page){
-        List<Sort.Order> sorts = new ArrayList<>();
-        sorts.add(Sort.Order.desc("createDate"));
-        Pageable pageable = PageRequest.of(page,10,Sort.by(sorts));
-        return questionRepository.findAll(pageable);
-    }
-
-    public void modify(Question question, String subject, String content) {
-        question.setSubject(subject);
-        question.setContent(content);
-        question.setModifyDate(LocalDateTime.now());
-        questionRepository.save(question);
-    }
-
-    public void delete(Question question){
-        questionRepository.delete(question);
-    }
-
-    public void vote(Question question, SiteUser siteUser){
-        question.getVoter().add(siteUser);
-        questionRepository.save(question);
-    }
-}
