--- src/main/java/org/example/user/user/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package org.example.user.user.repository;
-
-import org.example.user.user.entity.SiteUser;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<SiteUser, Long> {
-    Optional<SiteUser> findByUsername(String username);
-}
