+++ src/main/resources/templates/question_form.html
@@ -12,6 +12,21 @@
             <label for="content" class="form-label">내용</label>
             <textarea th:field="*{content}" class="form-control" rows="10"></textarea>
         </div>
+        <div class="mb-3">
+            <label for="paid" class="form-label">유료글 설정</label>
+            <div class="form-check">
+                <input class="form-check-input" type="radio" name="paid" id="paidTrue" value="true">
+                <label class="form-check-label" for="paidTrue">
+                    유료
+                </label>
+            </div>
+            <div class="form-check">
+                <input class="form-check-input" type="radio" name="paid" id="paidFalse" value="false" checked>
+                <label class="form-check-label" for="paidFalse">
+                    무료
+                </label>
+            </div>
+        </div>
         <input type="submit" value="저장하기" class="btn btn-primary my-2">
     </form>
 </div>
