+++ src/test/java/org/example/question/question/repository/QuestionRepositoryTest.java
@@ -1,6 +1,7 @@
 package org.example.question.question.repository;
 
-import org.example.question.question.entity.Question;
+import org.example.domain.question.question.entity.Question;
+import org.example.domain.question.question.repository.QuestionRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
